### 2018-07-04 e0e060f
Update paper.md
