### 2018-07-04 #805 @arfon
@sdtaylor - https://github.com/sdtaylor/pyPhenology/pull/87 should fix your paper.
