### 2018-07-04 bd9ddc4
Update paper.md (#87)
