### 2018-07-04 #805 @molgor
Hi @sdtaylor I´d be interested but I'm new to this. Do you have any guideline?
