### 2018-07-05 #805 @sdtaylor
Reviewer guidelines are here http://joss.theoj.org/about#reviewer_guidelines, hopefully an editor will chime in with more details.
