### 2018-07-05 #805 @Chilipp
Hi @sdtaylor and @arfon : I am not an expert on phenology, to be honest, so I cannot judge the usefulness, comprehensiveness and correct implementation of the software. Here, @jsta would be a better reviewer, probably.
But if that is no problem for you (and for @arfon), I am happy to contribute here by reviewing and testing the technical aspects of pyPhenology and work myself through the docs.
