### 2018-07-06 #805 @arfon
> But if that is no problem for you (and for @arfon), I am happy to contribute here by reviewing and testing the technical aspects of pyPhenology and working myself through the docs.

Hi @molgor - as @sdtaylor mentioned above, we post our review guidelines here: http://joss.theoj.org/about#reviewer_guidelines

I'm also working on an updated set of docs over in https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .

Basically, the review happens in a GitHub issue, is checklist-driven, and is generally more of a conversation with the author (and other reviewer sometime). If this sounds like something you might be able to help with I'll go ahead and start setting up the review thread.
