### 2018-07-13 364afea
add score method to bootstrap model (#90)
also some tests for this for both ensembles. #78
