### 2018-07-18 #805 @pjotrp
@assign @Chilipp as reviewer
