### 2018-07-19 #827 @whedon
[REVIEW]: pyPhenology: A python framework for plant phenology modelling
**Submitting author:** @sdtaylor (<a href="http://orcid.org/0000-0002-6178-6903">Shawn Taylor</a>)
**Repository:** <a href="https://github.com/sdtaylor/pyPhenology" target ="_blank">https://github.com/sdtaylor/pyPhenology</a>
**Version:** v0.5.0
