### 2018-07-19 #827 @pjotrp
@Chilipp and @harmn we are starting review on this issue tracker.

@sdtaylor to expedite the review process do you mind going through above list of check boxes and make sure they can be ticked (you can't tick them). Also check the PDF output carefully. Ping us here when you are done.
