+++ pyPhenology/models/__init__.py
@@ -1,7 +1,7 @@
 from .thermaltime import ThermalTime, M1, FallCooling
 from .chuine_models import Uniforc, Unichill
 from .alternating import Alternating, MSB
-from .ensemble_models import BootstrapModel, WeightedEnsemble
+from .ensemble_models import BootstrapModel, WeightedEnsemble, Ensemble
 from .stat_models import Linear, Naive
 from .sequential import Sequential
 from .example_model import Model
@@ -17,5 +17,6 @@ __all__ = ['ThermalTime',
            'M1',
            'Naive',
            'WeightedEnsemble',
+           'Ensemble',
            'FallCooling',
            'Model']
