+++ pyPhenology/models/ensemble_models.py
@@ -222,6 +222,183 @@ class BootstrapModel():
                                      model_file=filename,
                                      overwrite=overwrite)
 
+class Ensemble():
+    """Fit an ensemble of different models.
+    
+    This model can fit multiple models into an ensemble where the 
+    weights are equal among all ensemble members.
+    
+    Note that the core models must be passed initialized. They will be
+    fit within the Ensemble model::
+
+                from pyPhenology import models, utils
+                observations, predictors = utils.load_test_data(name='vaccinium')
+                
+                m1 = models.Thermaltime(parameters={'T':0})
+                m2 = models.Thermaltime(parameters={'T':5})
+                m3 = models.Thermaltime(parameters={'T':-5})
+                m4 = models.Thermaltime(parameters={'T':10})
+                m5 = models.Uniforc(parameters={'t1':1})
+                m6 = models.Uniforc(parameters={'t1':30})
+                m7 = models.Uniforc(parameters={'t1':60})
+                
+                ensemble = models.Ensemble(core_models=[m1,m2,m3,m4,m5,m6,m7])
+                ensemble.fit(observations, predictors)
+    
+    """
+    def __init__(self, core_models):
+        """Ensemble model
+        
+        core_models : list of pyPhenology models, or a saved model file
+
+        """
+        self.observations = None
+        self.predictors = None
+
+        if isinstance(core_models, list):
+            # List of models to fit
+            self.model_list = core_models
+            self.weights = np.array([None] * len(core_models))
+            
+        elif isinstance(core_models, str):
+            # A filename pointing toward a file from save_params()
+            model_info = utils.misc.read_saved_model(model_file=core_models)
+
+            if type(self).__name__ != model_info['model_name']:
+                raise RuntimeError('Saved model file does not match model class. ',
+                                   'Saved file is {a}, this model is {b}'.format(a=model_info['model_name'],
+                                                                                 b=type(self).__name__))
+
+            self.model_list = []
+            for model in model_info['core_models']:
+                Model = load_model(model['model_name'])
+                self.model_list.append(Model(parameters=model['parameters']))
+
+        else:
+            raise TypeError('core_models must be list of pyPhenology models',
+                            'or a filename for a saved model')
+    
+    def fit(self, observations, predictors, verbose=False, debug=False, **kwargs):
+        """Fit the underlying core models
+
+        Parameters:
+            observations : dataframe
+                pandas dataframe of phenology observations
+
+            predictors : dataframe
+                pandas dataframe of associated predictors
+
+            kwargs :
+                Other arguments passed to core model fitting (eg. optimzer methods)
+        """
+        self.observations = observations
+        self.predictors = predictors
+
+        for model in self.model_list:
+            model.fit(observations, predictors, **kwargs)
+
+    def predict(self, to_predict=None, predictors=None, aggregation='mean', **kwargs):
+        """Make predictions..
+
+        Predictions will be made using each core models, then a final prediction
+        for each observation using the specified aggregation method.
+
+        Parameters:
+            see core model description
+            
+            aggregation : str
+                Either 'mean', 'median', or 'none'. If using 'none' this returns
+                an array of tuple of size (number of members, predictions). 
+        """
+        if predictors is None:
+            predictors = self.predictors
+        if to_predict is None:
+            to_predict = self.observations
+
+        predictions = []
+        for model in self.model_list:
+            predictions.append(model.predict(to_predict=to_predict,
+                                             predictors=predictors,
+                                             **kwargs))
+
+        predictions = np.array(predictions)
+        if aggregation == 'mean':
+            predictions = np.mean(predictions, 0)
[…]
