### 2018-07-23 27a0b1e
add flexability in model loading
Make fitted  model loading more general so that it's easier
for load_saved_model or other ensemble methods to load
these ensemble models
