+++ pyPhenology/models/ensemble_models.py
@@ -52,24 +52,26 @@ class BootstrapModel():
             # A filename pointing toward a file from save_params()
             # core_model and num_bootstraps is ignored
             model_info = utils.misc.read_saved_model(model_file=parameters)
-
-            if type(self).__name__ != model_info['model_name']:
-                raise RuntimeError('Saved model file does not match model class. ' +
-                                   'Saved file is {a}, this model is {b}'.format(a=model_info['model_name'],
-                                                                                 b=type(self).__name__))
-
-            for bootstrap_iteration in model_info['parameters']:
-                Model = load_model(bootstrap_iteration['model_name'])
-                self.model_list.append(Model(parameters=bootstrap_iteration['parameters']))
-
+            self._parse_fully_fitted_model(model_info)
+ 
         elif isinstance(parameters, dict):
-            # Custom parameter values to pass to each bootstrap model
-            if core_model is None or num_bootstraps is None:
-                raise TypeError('core_model and num_bootstraps must be set')
-
-            validation.validate_model(core_model())
-            for i in range(num_bootstraps):
-                self.model_list.append(core_model(parameters=parameters))
+            # A dictionary passed in the parameters argument can either be
+            # a list of parameters to pass to the core model (such as in the 
+            # example above), OR a full fitted model specification (essentially
+            # the read json from a saved model file). The latter is only used 
+            # if this is from another ensemble method
+            
+            if 'model_name' in parameters:
+                # A saved model
+                self._parse_fully_fitted_model(parameters)
+            else:
+                # Custom parameter values to pass to each bootstrap model
+                if core_model is None or num_bootstraps is None:
+                    raise TypeError('core_model and num_bootstraps must be set')
+    
+                validation.validate_model(core_model())
+                for i in range(num_bootstraps):
+                    self.model_list.append(core_model(parameters=parameters))
 
         elif isinstance(parameters, list):
             # If its the output of BootstrapModel.get_params()
@@ -78,6 +80,17 @@ class BootstrapModel():
                 self.model_list.append(core_model(parameters=bootstrap_iteration))
         else:
             raise TypeError('parameters must be str or dict, got: ' + str(type(parameters)))
+
+    def _parse_fully_fitted_model(self, model_info):
+        # This loads a model from a saved json file
+        if type(self).__name__ != model_info['model_name']:
+            raise RuntimeError('Saved model file does not match model class. ' +
+                               'Saved file is {a}, this model is {b}'.format(a=model_info['model_name'],
+                                                                             b=type(self).__name__))
+
+        for bootstrap_iteration in model_info['parameters']:
+            Model = load_model(bootstrap_iteration['model_name'])
+            self.model_list.append(Model(parameters=bootstrap_iteration['parameters']))
 
     def fit(self, observations, predictors, **kwargs):
         """Fit the underlying core models
@@ -236,13 +249,10 @@ class Ensemble():
                 
                 m1 = models.Thermaltime(parameters={'T':0})
                 m2 = models.Thermaltime(parameters={'T':5})
-                m3 = models.Thermaltime(parameters={'T':-5})
-                m4 = models.Thermaltime(parameters={'T':10})
-                m5 = models.Uniforc(parameters={'t1':1})
-                m6 = models.Uniforc(parameters={'t1':30})
-                m7 = models.Uniforc(parameters={'t1':60})
-                
-                ensemble = models.Ensemble(core_models=[m1,m2,m3,m4,m5,m6,m7])
+                m3 = models.Uniforc(parameters={'t1':1})
+                m4 = models.Uniforc(parameters={'t1':30})
+
+                ensemble = models.Ensemble(core_models=[m1,m2,m3,m4])
[…]
