+++ pyPhenology/utils.py
@@ -124,11 +124,11 @@ def load_saved_model(filename):
 
     # These ensemble methods have their own code for loading saved files
     if model_info['model_name'] == 'BootstrapModel':
-        model = models.BootstrapModel(parameters=filename)
+        model = models.BootstrapModel(parameters=model_info)
     elif model_info['model_name'] == 'WeightedEnsemble':
-        model = models.WeightedEnsemble(core_models=filename)
+        model = models.WeightedEnsemble(core_models=model_info)
     elif model_info['model_name'] == 'Ensemble':
-        model = models.Ensemble(core_models=filename)
+        model = models.Ensemble(core_models=model_info)
     else:
         # For all other ones just need to pass the parameters
         Model = load_model(model_info['model_name'])
