### 2018-07-23 2efbc45
standardize model saving routines
_get_model_info is now used throughout and returns
a dictionary which describes the model. This allows for better
nesting of models when with an ensemble of bootstraps
