### 2018-07-23 e6e43a3
more ensemble tests
these functional tests ensures that the saved model is being
loaded correctly
