### 2018-07-23 d04f114
ensemble aggregation test
functional test where the predictions should pretty much
always differ slightly
