### 2018-07-23 66ae3d4
combine ensemble tests
this combines all the tests for ensemble, bootstrap, and
weighted ensemble (and combintations of those)
