### 2018-07-23 3898e26
Equal weight ensemble (#94)
A simpler ensemble which gives equal weight to all members.

Some adjusting of how ensemble parameters are saved and loaded was needed to allow, for example, an ensemble of bootstrapped models.

Tests were added to account for new things, and also all ensemble model tests were consolidated into a single file.
