+++ pyPhenology/models/ensemble_models.py
@@ -1,6 +1,6 @@
 import numpy as np
 from . import utils
-from ..utils import load_model
+from ..utils import load_model, load_model_parameters
 from . import validation
 from copy import deepcopy
 import warnings
@@ -52,32 +52,45 @@ class BootstrapModel():
             # A filename pointing toward a file from save_params()
             # core_model and num_bootstraps is ignored
             model_info = utils.misc.read_saved_model(model_file=parameters)
-
-            if type(self).__name__ != model_info['model_name']:
-                raise RuntimeError('Saved model file does not match model class. ' +
-                                   'Saved file is {a}, this model is {b}'.format(a=model_info['model_name'],
-                                                                                 b=type(self).__name__))
-
-            for bootstrap_iteration in model_info['parameters']:
-                Model = load_model(bootstrap_iteration['model_name'])
-                self.model_list.append(Model(parameters=bootstrap_iteration['parameters']))
-
+            self._parse_fully_fitted_model(model_info)
+ 
         elif isinstance(parameters, dict):
-            # Custom parameter values to pass to each bootstrap model
-            if core_model is None or num_bootstraps is None:
-                raise TypeError('core_model and num_bootstraps must be set')
-
-            validation.validate_model(core_model())
-            for i in range(num_bootstraps):
-                self.model_list.append(core_model(parameters=parameters))
+            # A dictionary passed in the parameters argument can either be
+            # a list of parameters to pass to the core model (such as in the 
+            # example above), OR a full fitted model specification (essentially
+            # the read json from a saved model file). The latter is only used 
+            # if this is from another ensemble method
+            
+            if 'model_name' in parameters:
+                # A saved model
+                self._parse_fully_fitted_model(parameters)
+            else:
+                # Custom parameter values to pass to each bootstrap model
+                if core_model is None or num_bootstraps is None:
+                    raise TypeError('core_model and num_bootstraps must be set')
+    
+                validation.validate_model(core_model())
+                for i in range(num_bootstraps):
+                    self.model_list.append(core_model(parameters=parameters))
 
         elif isinstance(parameters, list):
-            # If its the output of BootstrapModel.get_params()
+            # If its the output of BootstrapModel._get_model_info()
             for bootstrap_iteration in parameters:
                 bootstrap_iteration.pop('bootstrap_num')
                 self.model_list.append(core_model(parameters=bootstrap_iteration))
         else:
             raise TypeError('parameters must be str or dict, got: ' + str(type(parameters)))
+
+    def _parse_fully_fitted_model(self, model_info):
+        # This loads a model from a saved json file
+        if type(self).__name__ != model_info['model_name']:
+            raise RuntimeError('Saved model file does not match model class. ' +
+                               'Saved file is {a}, this model is {b}'.format(a=model_info['model_name'],
+                                                                             b=type(self).__name__))
+
+        for bootstrap_iteration in model_info['parameters']:
+            fitted_bootstrap_iteration = load_model_parameters(bootstrap_iteration)
+            self.model_list.append(fitted_bootstrap_iteration)
 
     def fit(self, observations, predictors, **kwargs):
         """Fit the underlying core models
@@ -182,13 +195,27 @@ class BootstrapModel():
         [m._check_parameter_completeness() for m in self.model_list]
 
     def get_params(self):
-
+        """This returns list of dictionaries with parameters of each bootstrap model
+        """
+        self._check_parameter_completeness()
+        
         all_params = []
         for i, model in enumerate(self.model_list):
             all_params.append(deepcopy(model.get_params()))
             all_params[-1].update({'bootstrap_num': i})
 
         return all_params
+
+    def _get_model_info(self):
+        # essentially the same as get_params() but is in a formate capable of
[…]
