+++ README.md
@@ -32,7 +32,7 @@ A Thermal Time growing degree day model:
 ```
 from pyPhenology import models, utils
 observations, predictors = utils.load_test_data(name='vaccinium')
-model = models.Thermal_Time()
+model = models.ThermalTime()
 model.fit(observations, predictors)
 model.get_params()
 {'t1': 85.704951490688927, 'T': 7.0814430573372666, 'F': 185.36866570243012}
@@ -41,7 +41,7 @@ model.get_params()
 Any of the parameters in a model can be set to a fixed value. For example the thermal time model with the threshold T set to 0 degrees C
 
 ```
-model = models.Thermal_Time(parameters={'T':0})
+model = models.ThermalTime(parameters={'T':0})
 model.fit(observations, predictors)
 model.get_params()
 {'t1': 26.369813953905265, 'F': 333.76534368004388, 'T': 0}
