### 2018-07-26 7c00909
minor doc updates
-emphasize model loading utility over other method
-missing newline in WeightedEnsemble
