### 2018-08-06 c61ec20
Fixed chuine2017 reference
Corrected DOI and Règni'ere
