### 2018-08-06 #827 @Chilipp
Hey! I really apologize for the late response, sorry.

I had a closer look at your software and enjoyed it a lot, thanks for your work here! To me it looks like a rather small package (only about ~1000 statements) but the editor knew this, when he accepted your software for review. So there are some minor points that need to be corrected from my point of view:

## General checks

- [ ] **Version**
> Does the release version given match the GitHub release (v0.5.0)?

Only a minor thing here: in https://github.com/sdtaylor/pyPhenology/releases there are two versions available: https://github.com/sdtaylor/pyPhenology/releases/tag/v0.5.0 and  https://github.com/sdtaylor/pyPhenology/releases/tag/0.5.0.
I assume, that v0.5.0 is the official release, so what about tag 0.5.0?

## Documentation

- [ ] **Installation instructions**
> Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

What about the different OS? To me, it looks like it runs on all platforms (Linux, MaxOSX and Windows) but it would be better to have written in the docs. This could be solved by merging https://github.com/sdtaylor/pyPhenology/pull/96

- [ ] **Example usage**
> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

You have a quickstart example in your docs (http://pyphenology.readthedocs.io/en/master/quickstart.html) but this is very short. Just be a bit more verbose in order to be more transparent, as well.
For example, add some statistics and/or plots on how the prediction compares against the observations (also in order to show the user, how he can compare the two.)

- [ ] **Functionality documentation**
> Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I like your documentation and your classes are well documented. Only the [Model.score](http://pyphenology.readthedocs.io/en/master/generated/pyPhenology.models.Model.score.html#pyPhenology.models.Model.score) method needs a better docstring but I have the impression that this method is not really maintained (and tested) anyway, since your are not using it nor testing it.

- [ ] **Automated tests**
> Are there automated tests or manual steps described so that the function of the software can be verified?

Your tests are fine and complete, however I strongly recommend to use a coverage service as well since this is super easy for a pure python package and adds more transparency to your open source project. Just use [pytest-cov](https://pytest-cov.readthedocs.io/en/latest/) on your CI and [codecov](https://codecov.io/) or [coveralls](https://coveralls.io/)

- [ ] **Community guidelines**
> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I could not find them, neither in the repository, nor in the docs. Please add the necessary file (e.g. a [CONTRIBUTING](https://help.github.com/articles/setting-guidelines-for-repository-contributors/) file in the main repository)

## Software paper

- [ ] **References**
> Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

1. Some DOI and names were wrong (see https://github.com/sdtaylor/pyPhenology/pull/97)
2. Is there any reference that you should provide for [the test data you use in your docs](file:///Users/psommer/Documents/reviews/pyPhenology/pyPhenology/docs/_build/html/utils.html#test-data)? Where is this data coming from?
