### 2018-08-07 #827 @pjotrp
@sdtaylor can you address these points?
