+++ README.md
@@ -1,7 +1,9 @@
 # pyPhenology  
 [![Build Status](https://travis-ci.org/sdtaylor/pyPhenology.svg?branch=master)](https://travis-ci.org/sdtaylor/pyPhenology) 
 [![License](http://img.shields.io/badge/license-MIT-blue.svg)](https://raw.githubusercontent.com/sdtaylor/phPhenology/master/LICENSE)
-[![Documentation Status](https://readthedocs.org/projects/pyphenology/badge/?version=master)](http://pyphenology.readthedocs.io/en/master/?badge=master)  
+[![Documentation Status](https://readthedocs.org/projects/pyphenology/badge/?version=master)](http://pyphenology.readthedocs.io/en/master/?badge=master)
+[![codecov](https://codecov.io/gh/sdtaylor/pyPhenology/branch/master/graph/badge.svg)](https://codecov.io/gh/sdtaylor/pyPhenology)  
+
 
 Plant phenology models in python with a scikit-learn inspired API
 
