### 2018-08-12 b662cca
add tests fore base.score()
