### 2018-08-12 90b1ce3
spruce up score method
add some more info to the docstring.
add some checks to make sure things are being passed correctly.
