### 2018-08-12 417998d
remove newline spaces
