### 2018-08-12 d28833a
fixed typo in license url (#95)
