+++ paper/paper.bib
@@ -68,7 +68,7 @@
   issn = {15452069},
   doi = {10.1146/annurev-ecolsys-110316-022706},
   journal = {Annu. Rev. Ecol. Evol. Syst},
-  author = {Chuine, Isabelle and R{\`e}gni'ere, Jacques},
+  author = {Chuine, Isabelle and R{\'e}gni{\`e}re, Jacques},
   year = {2017},
   keywords = {annual cycle,diapause,dormancy,environmental control,extratropical species,parameter estimation,response functions},
   pages = {159-82},
