### 2018-08-12 8dd059f
Fixed chuine2017 reference (#97)
* Fixed chuine2017 reference

Corrected DOI and Régnière
