### 2018-08-12 f70516a
add data citations to joss paper
