+++ paper/paper.bib
@@ -90,4 +90,25 @@
   pages = {1276-1285}
 }
 
+@Misc{npndata,
+  Author = {{USA National Phenology Network}},
+  Howpublished = {USA-NPN, Tucson, Arizona, USA. Data set accessed 05/01/2017 at \url{http://doi.org/10.5066/F78S4N1V}},
+  Title = {Plant and Animal Phenology Data. Data type: Status and Intensity. 01/01/2009-04/31/2017 for Region: 49.9375°, -66.4791667° (UR); 24.0625°, -125.0208333° (LL).},
+  Year = {2017},
+}
 
+@Misc{harvarddata,
+  Author = {O'Keefe, J.},
+  Title = {Phenology of Woody Species at Harvard Forest since 1990. Harvard Forest Data Archive: HF003},
+  Year = {2015},
+  Howpublished = {Data set accessed 05/01/2017 at \url{http://harvardforest.fas.harvard.edu:8080/exist/apps/datasets/showData.html?id=hf003}},
+}
+
+@Misc{prismdata,
+  Author = {{PRISM Climate Group}},
+  Howpublished = {\url{http://prism.oregonstate.edu}},
+  Publisher = {Oregon State University},
+  Title = {Oregon State University},
+  Year = {2004},
+  Url = {http://prism.oregonsate.edu}
+}
