+++ paper/paper.bib
@@ -108,7 +108,5 @@
   Author = {{PRISM Climate Group}},
   Howpublished = {\url{http://prism.oregonstate.edu}},
   Publisher = {Oregon State University},
-  Title = {Oregon State University},
-  Year = {2004},
-  Url = {http://prism.oregonsate.edu}
+  Year = {2004}
 }
