### 2018-08-12 bd7391b
Be a bit more verbose about supported OS (#96)
