### 2018-08-13 576fd99
expand the quickstart to be more detailed
-also bring in the autosection label module for easier referencing
-slight adjustments to the optimizer page to make things unique
