### 2018-08-13 ef3d12f
point to the examples at the end of the quickstart
