### 2018-08-13 ae2f44c
fix model details
