+++ docs/quickstart.rst
@@ -4,41 +4,107 @@
 Quickstart
 ====================
 
-A Thermal Time Model
---------------------
+Fitting a Thermal Time Model
+----------------------------
 
-An example dataset of blueberry (Vaccinium corymbosum) flower and leaf phenology from Harvard forest is available.::
+This short overview will fit a basic :any:`ThermalTime` model, which has 3 parameters.
+We will fit the model using data of flowering observations of blueberry (Vaccinium corymbosum)
+from the Harvard Forest, which can be loaded from the package::
 
     from pyPhenology import models, utils
-    observations, predictors = utils.load_test_data(name='vaccinium')
+    import numpy as np
+
+    observations, predictors = utils.load_test_data(name='vaccinium', phenophase='flowers')
 
 The two objects (observations and predictors) are pandas ``data.frames``. The observations ``data.frame`` contains the direct
-phenology observations as well as an associated site and year for each. predictors ``data.frame`` has several
-variables used to predict phenology, such as daily temperature measurements, latitude and longitude, and daylength.
-They listed for each site and year represented in observations. Both of these ``data.frames`` are required for building models.
+phenology observations as well as an associated site and year for each. The predictors ``data.frame`` has several
+variables used to predict phenology, such as daily mean temperature, latitude and longitude, and daylength.
+They are listed for each site and year represented in observations. Both of these ``data.frames`` are required for building models.
 Read more about how phenology data is structured in this package :ref:`here <data_structure>`.
 
-Initialize and fit a :any:`ThermalTime` model, which has 3 parameters::
+Next load the model::
 
     model = models.ThermalTime()
+
+Then fit the model using the vaccinium flowering data::
+
     model.fit(observations, predictors)
+
+The models estimated parameters can be viewed via a method call::
+
     model.get_params()
+    {'F': 356.9379498434921, 'T': 5.507956473407982, 't1': 27.33480003499163}
 
-A model can also be loaded via a text string with :any:`load_model`::
+Here, the fitted vaccinium flower model has a temperature threshold ``T`` of
+5.5 degrees C, a starting accumulation day ``t1`` of julian day 27, and total degree day
+requirement ``F`` of 357 units.
 
-    Model = utils.load_model('ThermalTime')
-    model = Model()
-    model.fit(observations, predictors)
-
-Using predict will give predictions of the same data as was used for fitting::
+The fitted model can also be used to predict the day of flowering based off the
+fitted parameters. By default it will make  predictions on the data used for
+fitting::
 
     model.predict()
+    array([126, 126, 127, 127, 126, 129, 129, 127, 132, 132, 133, 133, 132,
+      132, 130, 130, 130, 129, 127, 126, 132, 130, 129, 132, 132, 133,
+      133, 137, 137, 141, 141, 142, 132, 141, 141, 139, 139, 139, 139,
+      137, 137, 141, 141, 141, 141, 142, 142, 142])
 
-New predictions can be made by passing a new observations and temp data.frames, where the `doy` column in
-observations is not required. For example here we fit the model and prediction on held out data::
+Note that model predictions are different than the actual observed predictions.
+For example the root mean square error of the the predictions is about 3 days::
 
-    observations_test = observations[1:10]
-    observations_train = observations[10:]
+    np.sqrt(((model.predict() - observations.doy)**2).mean())
+    2.846781808756454
 
-    model.fit(observations_train, predictors)
-    observations_test['prediction_doy'] = model.predict(observations_test, predictors)
+This value can also be calculated by the model directly using :any:`Model.score`::
+
[…]
