+++ docs/optimizers.rst
@@ -2,7 +2,7 @@
 Optimizer Methods
 #################
 
-Parameters of phenology models have a complex search space and are commonly fit with `global optimization <https://en.wikipedia.org/wiki/Global_optimization>`__ algorithms. 
+Parameters of phenology models have a complex search space and are commonly fit with `global optimization <https://en.wikipedia.org/wiki/Global_optimization>`__ algorithms.
 To estimate parameters pyPhenology uses optimizers built-in to `scipy <https://docs.scipy.org/doc/scipy/reference/optimize.html#global-optimization>`__.
 Optimizers available are:
 
@@ -17,7 +17,7 @@ The optimizer can be specified in the ``fit`` method by using the codes ``DE``, 
 
     from pyPhenology import models, utils
     observations, temp = utils.load_test_data(name='vaccinium')
-    
+
     model = models.ThermalTime(parameters={})
     model.fit(observations, temp, method='DE')
 
@@ -27,13 +27,13 @@ Optimizer Arguments
 
 Arguments to the optimization algorithm are crucial to model fitting. These control things like the maximimum number of iterations and how to specify convergence.
 Ideally one should choose arguments which find the "true" solution, yet this is a tradeoff of time and resources available. Models with a large number of parameters (such as the Unichill model) can take
-several hours to days to fit if the optimizer arguments are set liberally. 
+several hours to days to fit if the optimizer arguments are set liberally.
 
 Optimizer arguments can be set two ways. The first is using some preset defaults::
 
     model.fit(observations, temp, method='DE', optimizer_params='practical')
 
-* ``testing``  Designed to be quick for testing code. Results from this should not be used for analysis. 
+* ``testing``  Designed to be quick for testing code. Results from this should not be used for analysis.
 * ``practical`` Default. Should produce realistic results on desktop systems in a relatively short period.
 * ``intensive`` Designed to find the absolute optimal solution. Can potentially take hours to days on large datasets.
 
@@ -42,8 +42,8 @@ The 2nd is using a dictionary for customized optimizer arguments::
 
     model.fit(observations, temp, method='DE',optimizer_params={'popsize': 50,
                                                                 'maxiter': 5000})
-                                                                
-                                                                
+
+
 All the arguments in the scipy optimizer functions are available via the ``optimizer_params`` argument in ``fit``. The important
 ones are described below, but also look at the available options in the scipy documentation. Any arguments not set will be
 set to the default specifed in the scipy package. Note that the three presets can be used with all optimizer methods, but using
@@ -57,18 +57,18 @@ Optimizers
 
 Differential Evolution
 ----------------------
-Differential evolution uses a population of models each randomly initialized to different parameter values within the respective search spaces. 
+Differential evolution uses a population of models each randomly initialized to different parameter values within the respective search spaces.
 Each "member" is adjusted slightly based on the performance of the best model. This process is repeated until the ``maxiter`` value is reached
-or a convergence threshold is met. 
+or a convergence threshold is met.
 
[…]
