+++ codecov.yml
@@ -1,0 +1,2 @@
+# Turn on commenting on PRs
+comment: false
