### 2018-08-13 #827 @sdtaylor
@pjotrp here is my reply for @Chilipp's review

Thanks for reviewing my package and your comments, I appreciate the your thoroughness. Though it is a relatively small package, the process of programming phenology models is a major bottleneck for most researchers, so I hope it will help them out in that regard. As detailed below I've implemented all of your suggestions.

## General Checks

#### Version
The 0.5.0 tag was created mistakenly. I've deleted it and now there is only v0.5.0

## Documentation

#### Installation instructions
pyPhenology does run on all operating systems. Thank you for PR in helping clarifying this, I've merged it.

#### Example usage
I've expanded the quickstart to better describe some of the common usage, such as parameter
adjustment and saving models. I also included a link at the end to point users
towards the examples which represent a more realistic analysis using this package.
You can see the changes [here](https://github.com/sdtaylor/pyPhenology/pull/99/files#diff-9d8a6f5b32d4a70a56734fd0eb50dd5b)

#### Functionality Documentation
Thanks for pointing this out. I would like the score method to be more prominent. I've
updated the docstring to be more clear and added some tests specifically for this here https://github.com/sdtaylor/pyPhenology/pull/98.

#### Automated tests
I have setup codecov for the repository and added the badge to the GitHub readme.

#### Community guidelines
I've added a [CONTRIBUTER.md](https://github.com/sdtaylor/pyPhenology/blob/master/CONTRIBUTING.md) file outlining the basics of contributing, as well
as a [CODE_OF_CONDCUCT.md](https://github.com/sdtaylor/pyPhenology/blob/master/CODE_OF_CONDUCT.md) file.

## Software Paper

#### References

1. I've merged this PR, thanks for fixing this DOI.
2. I've included citations for the 3 datasets used as well as a short description of them in the paper.
