### 2018-08-14 #827 @pjotrp
@Chilipp can you check and tick the relevant boxes? If you are happy we can wrap up. @harmn I assume you are happy about this submission?
