### 2018-08-14 #827 @harmn
@pjotr, only two very minor issues:
-the minimal required version for pandas (0.18.0) seems to be incorrect,Series does not have an isna method until 0.21 (used in pyPhenology/models/utils/misc.py on line 53)
-the Basler reference in the text has additional parentheses around the year
