### 2018-08-14 38cd200
pandas 0.21 required for Seriesna() use
