+++ requirements.txt
@@ -1,3 +1,3 @@
 scipy>=1.0 --only-binary :all:
 numpy>=1.8 --only-binary :all:
-pandas>=0.18 --only-binary :all:
+pandas>=0.21 --only-binary :all:
