### 2018-08-14 #827 @sdtaylor
Thank you @harmn for reviewing my package. I tested it with pandas 0.18.0 and you are correct that it doesn't work due to this incomparability. I've changed the pandas requirement to 0.21.0.

For the reference issue, I tried changing the ordering from `[@tang2016, @basler2016]` to `[@basler2016, @tang2016]` and now there is parenthesis around the year of the Tang reference. I'm not sure whats causing that, @pjotrp do you have any ideas?
