### 2018-08-14 #827 @Chilipp
@pjotrp  and @sdtaylor:

> @Chilipp can you check and tick the relevant boxes? If you are happy we can wrap up. @harmn I assume you are happy about this submission?

Done! It's all fine from my point of view. So if The two points by @harmn are fixed, this paper should be accepted.
