### 2018-08-15 #827 @pjotrp
@arfon we are ready to accept this submission. Note there is a small issue with generating the PDF (two boxes up on this page). Great work everyone!
