### 2018-08-15 #827 @sdtaylor
Thanks @pjotrp, The DOI's all look good to me. The current release is no v0.6.0, and I've deposited at zenodo at https://doi.org/10.5281/zenodo.1345528
