### 2018-08-15 #827 @arfon
Looks like the citations need to be separated by semicolons in the `paper.md`: http://pandoc.org/MANUAL.html#citations :

> Citations go inside square brackets and are separated by semicolons.

I've made this change locally to the `paper.md`.
